Add files via upload
Here is the pattern for the face mask.  Start with four rectangles 6 inches by 7 inches, stack them, and pin them together with the cut out pattern on top.  Cut out along the outer dark line.  Then follow the instructions on the pattern or watch my YouTube Video.
You can use Adobe Reader to print the PDF -- make sure you check the "Actual Size" box to keep Adobe from shrinking the image to "fit the page".  When it's printed correctly, the diagram of the metal strip should be 6.25 inches long.